Raycasting support

+++ include/gui.h
@@ -7,7 +7,7 @@ void initialize_gui();
 void load_bitmap_font(char *path);
 int bitmap_index(char c);
 
-void render_text_bitmap(char *text, int x, int y, double size);
-void draw_button(char *text, int x, int y);
-void draw_dialog_box(char *text, int x, int y);
-void draw_meter(char *text, int x, int y, color c, double full);
+void render_text_bitmap(char *text, double x, double y, double size);
+void draw_button(char *text, double x, double y);
+void draw_dialog_box(char *text, double x, double y);
+void draw_meter(char *text, double x, double y, color c, double full);
